+++ src/wrappers/hidingView/component.tsx
@@ -8,16 +8,19 @@ import { wrappingComponent } from '../wrappingComponent';
 
 interface IStyledHidingViewProps extends IWrapperProps {
   isHidden?: boolean;
+  isInvisible?: boolean;
 }
 
 const StyledHidingView = wrappingComponent((Component: React.ComponentType<IStyledHidingViewProps>): React.ComponentType<IStyledHidingViewProps> => {
   return styled(Component)<IStyledHidingViewProps>`
     ${(props: IStyledHidingViewProps): string => (props.isHidden ? 'display: none !important;' : '')};
+    ${(props: IStyledHidingViewProps): string => (props.isInvisible ? 'visibility: hidden !important;' : '')};
   `;
 });
 
 export interface IHidingViewProps extends IWrapperProps {
   isHidden?: boolean;
+  isInvisible?: boolean;
 }
 
 export const HidingView = (props: IHidingViewProps): React.ReactElement => {
@@ -25,6 +28,7 @@ export const HidingView = (props: IHidingViewProps): React.ReactElement => {
     <StyledHidingView
       className={getClassName(HidingView.displayName, props.className)}
       isHidden={props.isHidden}
+      isInvisible={props.isInvisible}
     >
       {props.children}
     </StyledHidingView>