+++ src/atoms/linkBase/component.tsx
@@ -91,9 +91,12 @@ export interface ILinkBaseProps extends IComponentProps<ILinkBaseTheme>, ISingle
 export const LinkBase = (props: ILinkBaseProps): React.ReactElement => {
   const isUsingCoreRouting = useIsCoreRoutingEnabled();
 
-  const onClicked = (): void => {
+  const onClicked = (event: React.SyntheticEvent): void => {
     if (props.onClicked) {
       props.onClicked();
+    }
+    if (props.onClicked || props.target) {
+      event.stopPropagation();
     }
   };