+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 -[MINOR] Fixed `Video` lazy loading
 -[MINOR] Make `Link` default style use inherit
 -[MINOR] Added `placeholderText` to `OptionSelect`
+-[MINOR] Use `optionTextVariant` correctly in `OptionSelect`
 -[MINOR] Fixed `alternateColors` to be expanded correctly
 
 ### Removed