+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 -[MINOR] Fixed `alternateColors` to be expanded correctly
 -[MINOR] Update `ColorSettingView` to set default `color` correctly
 -[MINOR] Added `isInvisible` to `HidingView`
+-[MINOR] Added `shouldBreakOnWords` and `shouldBreakAnywhere` to `Text`
 
 ### Removed