+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 -[MINOR] Added `shouldBreakOnWords` and `shouldBreakAnywhere` to `Text`
 -[MAJOR] Updated `Image` to support responsive images from pablo-images.kibalabs.com
 -[MINOR] Fixed `StatefulTitleCollapsibleBox` to pass through className
+-[MINOR] Updated `Button`, `IconButton`, `LinkBase` and `Link` to only show focus when navigating with keyboard
 
 ### Removed